Add Relations

+++ src/main/java/com/calcaddpayschoolbackend/entity/Result.java
@@ -28,7 +28,7 @@ public class Result {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "time_sheet_id")
     private TimeSheet timeSheet;
-    private int percent;
+    private double percent;
     private BigDecimal sum;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "basic_norms_id")
